Update system refactored
The update system has been refactored to enhance performance and eliminate dependency on the GitHub API.
