+++ app/templates/index.html
@@ -53,7 +53,7 @@
                value="Upload">
     </form>
     {% if session['update_needed'] == True %}
-        <a href="https://github.com/aviolaris/instaunfollowers/releases" class="blink">Update available!</a>
+        <a href="https://github.com/aviolaris/instaunfollowers/releases" class="blink"><b>Update available!</b></a>
     {% endif %}
     <div class="modal" tabindex="-1" role="dialog" id="errorModal">
         <div class="modal-dialog" role="document">
